Processors: Tons of changes. First of all, got rid of unnecessary JSON objects copying. Improved WT a bit (but is still bad). Made Delay, Filtering and TubeDist 8-channel - now everything is tied to AVX register size.
Next step should be improving NoteManager, getting rid of unnecessary math operations, etc.

+++ filters/filtering.hpp
@@ -49,7 +49,7 @@ public:
 	void process_callback() override;
 private:
 	void setup_filtering();
-	double freq = 0.0;
+	float8 freq = 0.0;
 	std::vector <std::unique_ptr <Dsp::Filter> > filters_bank;
 	Dsp::Params f_params;
